glean-clang: support deriving in addition to indexing
This patch also automates dumping the inventory and loading the indexed data,
using the glean library, and this can all be performed against a local or remote
glean database.

Using the same example C++ project as in 7945ad2, I tested this patch by doing:

``` sh
$ mkdir /gleandb
$ glean --db-root /gleandb \
        --db-schema dir:/Glean/glean/schema/source \
        create --repo foo/0
$ cabal run glean-clang:glean-clang -- --srcdir tmp/ --db-root /gleandb --schema dir:/Glean/glean/schema/source --repo foo/0
```

and then inspecting the resulting database in `glean shell`, making sure the
"raw" facts as well as the derived ones were there.